+++ __tests__/sequencerProvider.test.ts
@@ -1,9 +1,13 @@
-import { Contract, Provider, SequencerProvider, stark } from '../src';
-import { ZERO } from '../src/constants';
-import { feeTransactionVersion } from '../src/utils/hash';
+import {
+  BlockNumber,
+  Contract,
+  GetBlockResponse,
+  Provider,
+  SequencerProvider,
+  stark,
+} from '../src';
 import { toBN } from '../src/utils/number';
 import { encodeShortString } from '../src/utils/shortString';
-import { fromCallsToExecuteCalldata } from '../src/utils/transaction';
 import {
   compiledErc20,
   compiledL1L2,
@@ -19,6 +23,15 @@ describeIfSequencer('SequencerProvider', () => {
   const account = getTestAccount(sequencerProvider);
   let customSequencerProvider: Provider;
   let exampleContractAddress: string;
+  let exampleBlock: GetBlockResponse;
+  let exampleBlockNumber: BlockNumber;
+  let exampleBlockHash: string;
+
+  beforeAll(async () => {
+    exampleBlock = await sequencerProvider.getBlock('latest');
+    exampleBlockHash = exampleBlock.block_hash;
+    exampleBlockNumber = exampleBlock.block_number;
+  });
 
   describe('Gateway specific methods', () => {
     let exampleTransactionHash: string;
@@ -48,36 +61,6 @@ describeIfSequencer('SequencerProvider', () => {
       const transactionTrace = await sequencerProvider.getTransactionTrace(exampleTransactionHash);
       // TODO test optional properties
       expect(transactionTrace).toHaveProperty('signature');
-    });
-
-    test('simulate transaction', async () => {
-      const call = {
-        contractAddress: exampleContractAddress,
-        entrypoint: 'transfer',
-        calldata: [exampleContractAddress, '10', '0'],
-      };
-      const calldata = fromCallsToExecuteCalldata([call]);
-      const nonce = toBN(await account.getNonce());
-      const version = toBN(feeTransactionVersion);
-      const chainId = await account.getChainId();
-      const signature = await account.signer.signTransaction([call], {
-        walletAddress: account.address,
-        nonce,
-        version,
-        maxFee: ZERO,
-        chainId,
-      });
-
-      const res = await sequencerProvider.simulateTransaction(
-        {
-          contractAddress: account.address,
-          calldata,
-          signature,
-        },
-        { version, nonce }
-      );
-      expect(res).toHaveProperty('trace');
-      expect(res).toHaveProperty('fee_estimation');
     });
 
     test('getCode() -> { bytecode }', async () => {
@@ -154,6 +137,28 @@ describeIfSequencer('SequencerProvider', () => {
       const [res] = result;
       expect(res.low).toStrictEqual(toBN(1000));
       expect(res).toStrictEqual(result.balance);
+    });
+  });
+
+  describe('getBlockTraces', () => {
+    test(`getBlockTraces(blockHash=${exampleBlockHash}, blockNumber=undefined)`, async () => {
+      const blockTraces = await sequencerProvider.getBlockTraces(exampleBlockHash);
+      expect(blockTraces).toHaveProperty('traces');
+      expect(blockTraces.traces[0]).toHaveProperty('validate_invocation');
+      expect(blockTraces.traces[0]).toHaveProperty('function_invocation');
+      expect(blockTraces.traces[0]).toHaveProperty('fee_transfer_invocation');
+      expect(blockTraces.traces[0]).toHaveProperty('signature');
+      expect(blockTraces.traces[0]).toHaveProperty('transaction_hash');
+    });
+
+    test(`getBlockTraces(blockHash=undefined, blockNumber=${exampleBlockNumber})`, async () => {
+      const blockTraces = await sequencerProvider.getBlockTraces(exampleBlockNumber);
+      expect(blockTraces).toHaveProperty('traces');
+      expect(blockTraces.traces[0]).toHaveProperty('validate_invocation');
+      expect(blockTraces.traces[0]).toHaveProperty('function_invocation');
+      expect(blockTraces.traces[0]).toHaveProperty('fee_transfer_invocation');
+      expect(blockTraces.traces[0]).toHaveProperty('signature');
+      expect(blockTraces.traces[0]).toHaveProperty('transaction_hash');
     });
   });
 });